Questionably reorganised code for compile time benefits.
